Úpdate

+++ visualize.py
@@ -74,12 +74,15 @@ def draw_pick_frames(
         rmat = np.array(frame["rotation"])
         assert tvec.shape == (3,)
         assert rmat.shape == (3, 3)
-
         rvec = cv2.Rodrigues(rmat)[0]
-
+      
         image_pts, _ = cv2.projectPoints(axes, rvec, tvec, cam_mtx, cam_dist)
-        image_pts = image_pts.squeeze(1).astype(np.int32)
+        # _, rvecs,tvecs = cv2.solveP3P(axes[np.array([3,0,1])], image_pts[np.array([3,0,1])], cam_mtx, cam_dist, 5)
+        # image_pts, _ = cv2.projectPoints(axes, rvecs[0], tvecs[0], cam_mtx, cam_dist)
+        image_pts = image_pts.squeeze(1).astype(int)
+
         origin_pt = image_pts[3]
+        
 
         cv2.line(image_bgr, origin_pt, image_pts[0], (255, 0, 0), 5)
         cv2.line(image_bgr, origin_pt, image_pts[1], (0, 255, 0), 5)
